ArXiv_Search_MCP/requirements.txt
pyyaml
ascii_colors
fastmcp
arxiv
pipmaster